+++ packages/remirror__react-renderer/src/handlers/index.ts
@@ -1,0 +1,2 @@
+export { CodeBlock } from './code-block';
+export { TextHandler } from './text';
