+++ packages/remirror__react-renderer/src/index.ts
@@ -1,1 +1,3 @@
-export { RemirrorRenderer } from './renderer';
+export * from './handlers';
+export { Doc, RemirrorRenderer } from './renderer';
+export type { MarkMap } from './types';
