+++ packages/remirror__react-renderer/src/types.ts
@@ -1,0 +1,3 @@
+import { ComponentType } from 'react';
+
+export type MarkMap = Partial<Record<string, string | ComponentType<any>>>;
